# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "kindleM API",
        "contact": {},
        "license": {},
        "version": "0.0.1"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
        "/chapters": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "get manga chapter list",
                "parameters": [
                    {
                        "type": "string",
                        "description": " ",
                        "name": "manga_url",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.MangaInfo"
                        }
                    }
                }
            }
        },
        "/check/chapter/{chapter_id}": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "check chapter all picture download again",
                "parameters": [
                    {
                        "type": "integer",
                        "description": " ",
                        "name": "chapter_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.PictureList"
                        }
                    }
                }
            }
        },
        "/download": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "download chapter list",
                "parameters": [
                    {
                        "description": " ",
                        "name": "download_list",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.ChapterRowList"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.ChapterList"
                        }
                    }
                }
            }
        },
        "/search/{query}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "search manga",
                "operationId": "search-manga",
                "parameters": [
                    {
                        "type": "string",
                        "description": " ",
                        "name": "query",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.Manga"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Chapter": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "created": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "link": {
                    "type": "string"
                },
                "manga_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                },
                "updated": {
                    "type": "string"
                }
            }
        },
        "model.ChapterList": {
            "type": "array",
            "items": {
                "type": "object",
                "properties": {
                    "count": {
                        "type": "integer"
                    },
                    "created": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "link": {
                        "type": "string"
                    },
                    "manga_id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "total": {
                        "type": "integer"
                    },
                    "updated": {
                        "type": "string"
                    }
                }
            }
        },
        "model.ChapterRow": {
            "type": "object",
            "properties": {
                "link": {
                    "type": "string"
                },
                "manga_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "model.ChapterRowList": {
            "type": "array",
            "items": {
                "type": "object",
                "properties": {
                    "link": {
                        "type": "string"
                    },
                    "manga_id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    }
                }
            }
        },
        "model.Manga": {
            "type": "object",
            "properties": {
                "alias": {
                    "type": "string"
                },
                "author": {
                    "type": "string"
                },
                "cover": {
                    "type": "string"
                },
                "created": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "intro": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "source": {
                    "type": "integer"
                },
                "updated": {
                    "type": "string"
                }
            }
        },
        "model.MangaInfo": {
            "type": "object",
            "properties": {
                "ChapterRowList": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "link": {
                                "type": "string"
                            },
                            "manga_id": {
                                "type": "integer"
                            },
                            "title": {
                                "type": "string"
                            }
                        }
                    }
                },
                "alias": {
                    "type": "string"
                },
                "author": {
                    "type": "string"
                },
                "cover": {
                    "type": "string"
                },
                "created": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "intro": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "source": {
                    "type": "integer"
                },
                "updated": {
                    "type": "string"
                }
            }
        },
        "model.Picture": {
            "type": "object",
            "properties": {
                "chapterId": {
                    "type": "integer"
                },
                "created": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "index": {
                    "type": "integer"
                },
                "mangaId": {
                    "type": "integer"
                },
                "referer": {
                    "type": "string"
                },
                "src": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "updated": {
                    "type": "string"
                }
            }
        },
        "model.PictureList": {
            "type": "array",
            "items": {
                "type": "object",
                "properties": {
                    "chapterId": {
                        "type": "integer"
                    },
                    "created": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "index": {
                        "type": "integer"
                    },
                    "mangaId": {
                        "type": "integer"
                    },
                    "referer": {
                        "type": "string"
                    },
                    "src": {
                        "type": "string"
                    },
                    "status": {
                        "type": "boolean"
                    },
                    "updated": {
                        "type": "string"
                    }
                }
            }
        }
    }
}